Updated nodes so that declaration nodes can also invalidate parent context
The problem was that when the reconciler changed or moved (or removed) a declaration node - the parent node's context was not marked as changed (invalidated). This caused changes to the DOM from JS / reconciler not to be reflected in the native output.

This commit adds a system where a node will mark its own context (or the parent untill we have a parent that has a context) as changed when there are DOM operations.

- Updated logging: Moved level/indentation to node (instead of drawing context) and added parent pointer in children. Removed context as parameter from printDebugInfo