Changed Views to be driven by ReactViewManager (#1208)
On Android we were using regular views and viewmanagers to drive our native views. To be able to take advantage of the React Native functionality on views the Skia Views should be backed by a React View Manager which will ensure some functionality that will work out of the box.

This fixes #1197 and pointer events (use example in 1197 to test), and also allows for setting backgroundColor as style for a Canvas.

All tests and examples have been tested and works.