Removed EmptyProp value from Point props
Static EmptyProp value was used to detect points that was not set instead of the regular pattern of using nullptrs when props arent set.

This commit fixes this by removing the EmptyProps value and throwing an exception in the processValue if it tries to process a value that is not a prop.