Fixed retain cycle on the registerValues callback
The DependencyManager receives a callback on the ctor comming
from the Canvas object for making connections with animated values
and view ids. This was not cleared correctly creating a cyclic
dependency that caused the dom to not be released correctly.

This commit fixes this by explicitly setting the callback to null when
the dependencyManager's remove method is called. This is called when
the container is unmounted from the reconciler.